+++ content/debug-init.js
@@ -10,7 +10,7 @@ if (typeof chrome !== 'undefined' && chrome.storage) {
     chrome.storage.local.get('aladinn_debug_mode', (res) => {
         window.__ALADINN_DEBUG__ = !!res.aladinn_debug_mode;
         if (window.VNPTConfig) window.VNPTConfig.DEBUG = !!res.aladinn_debug_mode;
-        window.postMessage({ type: 'ALADINN_SET_DEBUG', state: !!res.aladinn_debug_mode }, '*');
+        window.postMessage({ type: 'ALADINN_SET_DEBUG', state: !!res.aladinn_debug_mode }, window.location.origin);
     });
 
     // Lắng nghe thay đổi từ storage